refactor(test): restructure change proof tests into change/ submodule
Move flat test files into a `change/` directory module for better
organization. Consolidate shared infrastructure (new_db, verify_and_check,
macros) and rewrite adversarial tests to use setup_db!/setup_2nd_commit!
macros, significantly reducing boilerplate. Remove 2 duplicate truncation
tests that existed in both change.rs and change_adversarial.rs.

Test count: 59 non-slow → 57 (minus 2 duplicates). 3 slow fuzz tests
unchanged.
