+++ firewood/src/merkle/tests/mod.rs
@@ -4,8 +4,6 @@
 #![expect(clippy::indexing_slicing, clippy::unwrap_used)]
 
 mod change;
-mod change_adversarial;
-mod changefuzz;
 #[cfg(feature = "ethhash")]
 mod ethhash;
 // TODO: get the hashes from merkledb and verify compatibility with branch factor 256
